Move shared code outwith the lambda folder
From an AWS perspective, the shared layer is accessed as if it's on the path. In this case, we can access `classes` as if it were on the same level as `index` - so we can do `from classes import CSMentor`.

This looks horrible in the IDE, because the IDE expects the path to be `from python.classes import...`

We get around this by telling the IDE that `python` is a "Source Folder" - right click and "Mark folder as"